Move ComponentPoint to separate file
