+++ packages/flame/lib/src/experimental/world.dart
@@ -4,6 +4,7 @@ import 'package:meta/meta.dart';
 import 'package:vector_math/vector_math_64.dart';
 
 import '../components/component.dart';
+import '../components/component_point.dart';
 import 'camera_component.dart';
 
 /// The root component for all game world elements.
